part5/requirements.txt
# Flask & Extensions
Flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.7.1
flask-restx==1.3.0
flask-migrate==4.1.0
flask-babel==4.0.0

# Database
SQLAlchemy==2.0.36

# Authentication
bcrypt==4.2.1
PyJWT==2.10.1

# Email (Resend)
resend==2.5.0

# Payment (Stripe)
stripe==11.4.1

# Maps & Geocoding
geopy==2.4.1

# Translation
googletrans==4.0.0rc1

# Media/Upload
Pillow==11.1.0

# Utilities
python-dotenv==1.0.1
marshmallow==3.23.2
gunicorn==23.0.0
requests==2.32.3

# Testing
pytest==8.3.4
pytest-cov==6.0.0
